Split test_discovery.py into per-module test files
Mirror the source package structure under tests/genai/discovery/:
- conftest.py: shared fixtures (make_trace, make_mock_span, make_assessment)
- test_utils.py: helper function tests
- test_pipeline.py: discover_issues integration tests
